+++ PianoFromAbove/MIDI.h
@@ -157,8 +157,8 @@ public:
     ~MIDITrack( void );
 
     //Parsing functions that load data into the instance
-    size_t ParseTrack( const unsigned char *pcData, size_t iMaxSize, int iTrack );
-    size_t ParseEvents( const unsigned char *pcData, size_t iMaxSize, int iTrack );
+    size_t ParseTrack( const unsigned char *pcData, size_t iMaxSize, size_t iTrack );
+    size_t ParseEvents( const unsigned char *pcData, size_t iMaxSize, size_t iTrack );
     void clear( void );
 
     friend class MIDIPos;
